+++ flink/src/test/java/io/delta/flink/utils/DeltaTableUpdater.java
@@ -1,12 +1,11 @@
-package io.delta.flink.source;
+package io.delta.flink.utils;
 
 import java.io.IOException;
 import java.util.Collections;
 import java.util.List;
 import java.util.UUID;
 
-import io.delta.flink.source.ContinuousTestDescriptor.Descriptor;
-import io.delta.flink.utils.DeltaTestUtils;
+import io.delta.flink.utils.ContinuousTestDescriptor.Descriptor;
 import org.apache.flink.api.common.serialization.BulkWriter;
 import org.apache.flink.core.fs.FileSystem.WriteMode;
 import org.apache.flink.core.fs.Path;
